+++ src/libs/TransactionRecoveryApi/constants/contract.ts
@@ -8,9 +8,11 @@ export const NATIVE_GAS_TOKEN_SYMBOL = {
   [EvmChain.FANTOM]: GasToken.FTM,
   [EvmChain.AURORA]: GasToken.AURORA,
   [EvmChain.BINANCE]: GasToken.BINANCE,
+  [EvmChain.BNBCHAIN]: GasToken.BNBCHAIN,
   [EvmChain.ARBITRUM]: GasToken.ETH,
   [EvmChain.CELO]: GasToken.CELO,
   [EvmChain.KAVA]: GasToken.KAVA,
+  [EvmChain.BASE]: GasToken.BASE,
 };
 
 export const DEFAULT_ESTIMATED_GAS = 700000;