+++ src/libs/TransactionRecoveryApi/helpers/index.ts
@@ -2,3 +2,4 @@ export * from "./axelarHelper";
 export * from "./contractEventHelper";
 export * from "./providerHelper";
 export * from "./contractCallHelper";
+export * from "./getCommandId";